Struct xterm_js_sys::crossterm_support::XtermJsCrosstermBackend

pub struct XtermJsCrosstermBackend<'a> {
    pub terminal: &'a Terminal,
    buffer: Cell<Vec<u8>>,
}

This is supported on feature="crossterm-support" only.

Wrapper for the xterm.js terminal for use with crossterm.

Fields
terminal: &'a Terminal

This is supported on feature="crossterm-support" only.

The xterm.js terminal that this struct instance wraps.

buffer: Cell<Vec<u8>>

This is supported on feature="crossterm-support" only.

Internal buffer for data to write to the terminal.

This lets us make one big call to Terminal::write with a batch of commands rather than many small calls.

Implementations

impl<'a> XtermJsCrosstermBackend<'a>

#[must_use]pub fn new(terminal: &'a Terminal) -> Self

This is supported on feature="crossterm-support" only.

Constructor for the wrapper type.

#[must_use]pub fn new_with_capacity(terminal: &'a Terminal, capacity: usize) -> Self

This is supported on feature="crossterm-support" only.

Like new except it also takes an estimate for the size of the internal buffer.

This is useful if you have a good guess about how many bytes the commands you're going to send will need.

pub fn write_immediately(&mut self, commands: String) -> IoResult<()>

This is supported on feature="crossterm-support" only.

Writes a String directly to the underlying terminal, bypassing the buffer.

This is useful for situations in which the commands being sent are already buffered and the extra copy is undesired. Note that this will flush the buffer first to preserve the order of commands.

Errors

This should never actually error. For consistency with the Write calls it results an io::Result

fn flush_immutable(&self) -> IoResult<()>

This is supported on feature="crossterm-support" only.

A version of flush that takes an immutable reference instead of a mutable one.

This exists because we want to flush the buffer in the Deref impl.