Module xterm_js_sys::ext

This is supported on feature="ext" only.

Glue for the Xterm.js types.

Re-exports

pub use addon::*;

pub use disposable::*;

pub use event::*;

pub use log_level::*;

pub use marker::*;

pub use options::*;

pub use theme::*;

pub use unicode::*;

Modules

addon	feature="ext"
Utilities for Addons.

disposable	feature="ext"
Utilities for Disposable.

event	feature="ext"
Sugar for methods that take IEvents.

log_level	feature="ext"
Conversions for LogLevel.

marker	feature="ext"
Extra functions for Markers.

options	feature="ext"
Helpers for TerminalOptions, Theme, and WindowOptions.

theme	feature="ext"
Some Themes.

unicode	feature="ext"
Helpers and duals for UnicodeVersionProvider and UnicodeHandling.

Macros

interface	feature="ext"
Creates a Rust trait to match a particular JS interface.

object	feature="ext"
Defines a JS object with some properties.

Traits

IntoJsInterface	feature="ext"
Represents a Rust type that satisfies a JS interface and can be turned into the concrete type that represents the JS interface.