Module xterm_js_sys::ext::disposable

This is supported on feature="ext" only.

Utilities for Disposable.

Structs

DisposableWrapper	feature="ext"
A wrapper for Disposable that calls dispose on Drop.

NoOpDispose	feature="ext"
A type that satisfies the Disposable interface and does nothing on dispose.

Traits

XtermDisposable	feature="ext"
Rust version of the Disposable interface.