Trait xterm_js_sys::ext::disposable::XtermDisposable

pub trait XtermDisposable {
    fn dispose(&self);

    fn into_js_by_ref(&self) -> Disposable
    where
        Self: Clone + 'static,
    { ... }

    fn into_js(self) -> Disposable
    where
        Self: Sized + 'static,
    { ... }

    fn into_js_inner(&'static self) -> Object
    where
        Self: 'static,
    { ... }
}

This is supported on feature="ext" only.

Rust version of the Disposable interface.

See the "mirroring interfaces" section of the xterm module docs for more information.

Required methods

fn dispose(&self)

This is supported on feature="ext" only.

Disposes of the instance.

This can involve unregistering an event listener or cleaning up resources or anything else that should happen when an instance is disposed of.

Dual of Disposable::dispose.

Loading content...
Provided methods

fn into_js_by_ref(&self) -> Disposable where
Self: Clone + 'static,

This is supported on feature="ext" only.

Copy of IntoJsInterface::by_ref.

fn into_js(self) -> Disposable where
Self: Sized + 'static,

This is supported on feature="ext" only.

Copy of IntoJsInterface::to.

fn into_js_inner(&'static self) -> Object where
Self: 'static,

This is supported on feature="ext" only.

Internal version of into_js_by_ref that doesn't leak self.

Useful for trait/interface hierarchies.

Loading content...