Struct xterm_js_sys::ReadOnlyArray

#[repr(transparent)]pub struct ReadOnlyArray<T: JsCast> {
    inner: Array,
    _t: PhantomData<T>,
}

Mirrors TypeScript's ReadOnlyArray interface.

This is built on the Array type in js-sys and primarily just mirrors over the subset of regular array functions that ReadOnlyArray provides. This wrapper also offers typed versions of all the functions that it forwards (prefixed with typed_).

Because wasm-bindgen doesn't support generic structs, this type uses #[repr(transparent)] and leans on its inner Array for all the wasm-bindgen specific impls.

Note that this struct does not provide an AsRef impl for Array as that would defeat the "read only" part of this type's guarantees.

Fields
inner: Array

The Array we're backed by.

_t: PhantomData<T>

Marker for T.

Implementations

impl<T: JsCast> ReadOnlyArray<T>

#[must_use]pub fn get(&self, index: u32) -> JsValue

Retrieves the element at the index (returns undefined if the index is out of range).

#[must_use]pub fn concat(&self, array: &Array) -> Array

The concat() method is used to merge two or more arrays. This method does not change the existing arrays, but instead returns a new array.

MDN documentation

#[must_use]pub fn every(
&self,
predicate: &mut dyn FnMut(JsValue, u32, Array) -> bool
) -> bool

The every() method tests whether all elements in the array pass the test implemented by the provided function.

MDN documentation

#[must_use]pub fn filter(
&self,
predicate: &mut dyn FnMut(JsValue, u32, Array) -> bool
) -> Array

The filter() method creates a new array with all elements that pass the test implemented by the provided function.

MDN documentation

#[must_use]pub fn find(
&self,
predicate: &mut dyn FnMut(JsValue, u32, Array) -> bool
) -> JsValue

The find() method returns the value of the first element in the array that satisfies the provided testing function. Otherwise undefined is returned.

MDN documentation

#[must_use]pub fn find_index(
&self,
predicate: &mut dyn FnMut(JsValue, u32, Array) -> bool
) -> i32

The findIndex() method returns the index of the first element in the array that satisfies the provided testing function. Otherwise -1 is returned.

MDN documentation

#[must_use]pub fn flat(&self, depth: i32) -> Array

The flat() method creates a new array with all sub-array elements concatenated into it recursively up to the specified depth.

MDN documentation

#[must_use]pub fn flat_map(
&self,
callback: &mut dyn FnMut(JsValue, u32, Array) -> Vec<JsValue>
) -> Array

The flatMap() method first maps each element using a mapping function, then flattens the result into a new array.

MDN documentation

pub fn for_each(&self, callback: &mut dyn FnMut(JsValue, u32, Array))

The forEach() method executes a provided function once for each array element.

MDN documentation

#[must_use]pub fn includes(&self, value: &JsValue, from_index: i32) -> bool

The includes() method determines whether an array includes a certain element, returning true or false as appropriate.

MDN documentation

#[must_use]pub fn index_of(&self, value: &JsValue, from_index: i32) -> i32

The indexOf() method returns the first index at which a given element can be found in the array, or -1 if it is not present.

MDN documentation

#[must_use]pub fn join(&self, delimiter: &str) -> JsString

The join() method joins all elements of an array (or an array-like object) into a string and returns this string.

MDN documentation

#[must_use]pub fn last_index_of(&self, value: &JsValue, from_index: i32) -> i32

The lastIndexOf() method returns the last index at which a given element can be found in the array, or -1 if it is not present. The array is searched backwards, starting at fromIndex.

MDN documentation

#[must_use]pub fn length(&self) -> u32

The length property of an object which is an instance of type Array sets or returns the number of elements in that array. The value is an unsigned, 32-bit integer that is always numerically greater than the highest index in the array.

MDN documentation

#[must_use]pub fn map(
&self,
predicate: &mut dyn FnMut(JsValue, u32, Array) -> JsValue
) -> Array

map() calls a provided callback function once for each element in an array, in order, and constructs a new array from the results. callback is invoked only for indexes of the array which have assigned values, including undefined. It is not called for missing elements of the array (that is, indexes that have never been set, which have been deleted or which have never been assigned a value).

MDN documentation

#[must_use]pub fn reduce(
&self,
predicate: &mut dyn FnMut(JsValue, JsValue, u32, Array) -> JsValue,
initial_value: &JsValue
) -> JsValue

The reduce() method applies a function against an accumulator and each element in the array (from left to right) to reduce it to a single value.

MDN documentation

#[must_use]pub fn reduce_right(
&self,
predicate: &mut dyn FnMut(JsValue, JsValue, u32, Array) -> JsValue,
initial_value: &JsValue
) -> JsValue

The reduceRight() method applies a function against an accumulator and each value of the array (from right-to-left) to reduce it to a single value.

MDN documentation

#[must_use]pub fn slice(&self, start: u32, end: u32) -> Array

The slice() method returns a shallow copy of a portion of an array into a new array object selected from begin to end (end not included). The original array will not be modified.

Note that because a copy is returned, the fact that an Array is produced from a ReadOnlyArray is not a problem.

MDN documentation

#[must_use]pub fn some(&self, predicate: &mut dyn FnMut(JsValue) -> bool) -> bool

The some() method tests whether at least one element in the array passes the test implemented by the provided function.

Note: This method returns false for any condition put on an empty array.

MDN documentation

#[must_use]pub fn to_locale_string(&self, locales: &JsValue, options: &JsValue) -> JsString

The toLocaleString() method returns a string representing the elements of the array. The elements are converted to Strings using their toLocaleString methods and these Strings are separated by a locale-specific String (such as a comma “,”).

MDN documentation

#[must_use]pub fn to_string(&self) -> JsString

The toString() method returns a string representing the specified array and its elements.

MDN documentation

#[must_use]pub fn iter(&self) -> ArrayIter<'_>

Returns an iterator over the values of the JS array.

#[must_use]pub fn to_vec(&self) -> Vec<JsValue>

Converts the JS array into a new Vec.

#[must_use]pub fn keys(&self) -> Iterator

The keys() method returns a new Array Iterator object that contains the keys for each index in the array.

MDN documentation

#[must_use]pub fn entries(&self) -> Iterator

The entries() method returns a new Array Iterator object that contains the key/value pairs for each index in the array.

MDN documentation

#[must_use]pub fn values(&self) -> Iterator

The values() method returns a new Array Iterator object that contains the values for each index in the array.

MDN documentation

#[must_use]pub fn is_array(value: &JsValue) -> bool

The Array.isArray() method determines whether the passed value is an Array.

MDN documentation

impl<T: JsCast> ReadOnlyArray<T>

#[must_use]pub fn typed_get(&self, index: u32) -> Option<T>

Typed version of Array::get.

Returns None when the element can't be casted into T or the index is out of bounds.

#[must_use]pub fn typed_concat(&self, array: &Self) -> Self

Typed version of Array::concat.

pub fn typed_every(&self, predicate: impl FnMut(T, u32, Self) -> bool) -> bool

Typed version of Array::every.

Returns false is any of the elements can't be casted into T.

#[must_use]pub fn typed_filter(&self, predicate: impl FnMut(T, u32, Self) -> bool) -> Self

Typed version of Array::filter.

Elements that can't be casted into T are automatically filtered out.

#[must_use]pub fn typed_find(
&self,
predicate: impl FnMut(T, u32, Self) -> bool
) -> Option<T>

Typed version of Array::find.

Elements that can't be casted into T are automatically ignored.

#[must_use]pub fn typed_find_index(
&self,
predicate: impl FnMut(T, u32, Self) -> bool
) -> Option<u32>

Typed version of Array::find_index.

Elements that can't be casted into T are automatically ignored.

Returns None when the element can't be found.

#[must_use]pub fn typed_flat(&self, depth: i32) -> Self

Typed version of Array::flat.

#[must_use]pub fn typed_flat_map(
&self,
callback: impl FnMut(T, u32, Self) -> Vec<T>
) -> Self

Typed version of Array::flat_map.

Elements that can't be casted into T are automatically ignored.

pub fn typed_for_each(&self, callback: impl FnMut(T, u32, Self))

Typed version of Array::for_each.

Elements that can't be casted into T are automatically ignored.

pub fn typed_includes(&self, value: T, from_index: i32) -> bool

Typed version of Array::includes.

Note that this takes the element to search for by value (we have no way to convert from &T to &JsValue); if T impls AsRef<JsValue>, you're better off using ReadOnlyArray::includes; it will be cheaper (but possibly less ergonomic).

pub fn typed_index_of(&self, value: T, from_index: i32) -> Option<u32>

Typed version of Array::index_of.

Note that this takes the element to search for by value (we have no way to convert from &T to &JsValue); if T impls AsRef<JsValue>, you're better off using ReadOnlyArray::index_of; it will be cheaper (but possibly less ergonomic).

Returns None when the element can't be found.

pub fn typed_last_index_of(&self, value: T, from_index: i32) -> Option<u32>

Typed version of Array::last_index_of.

Note that this takes the element to search for by value (we have no way to convert from &T to &JsValue); if T impls AsRef<JsValue>, you're better off using ReadOnlyArray::last_index_of; it will be cheaper (but possibly less ergonomic).

Returns None when the element can't be found.

#[must_use]pub fn typed_map<R: JsCast>(
&self,
predicate: impl FnMut(T, u32, Self) -> R
) -> ReadOnlyArray<R>

Typed version of Array::map.

Elements that can't be cast into T are mapped to JsValue::UNDEFINED.

#[must_use]pub fn typed_reduce<A: JsCast>(
&self,
predicate: impl FnMut(A, T, u32, Self) -> A,
initial_value: A
) -> A

Typed version of Array::reduce.

Effectively 'skips' elements that can't be cast to T (just returns the accumulator as is for those elements).

#[must_use]pub fn typed_reduce_right<A: JsCast>(
&self,
predicate: impl FnMut(A, T, u32, Self) -> A,
initial_value: A
) -> A

Typed version of Array::reduce_right.

Effectively 'skips' elements that can't be cast to T (just returns the accumulator as is for those elements).

#[must_use]pub fn typed_slice(&self, start: u32, end: u32) -> Self

Typed version of Array::slice.

#[must_use]pub fn typed_some(&self, predicate: impl FnMut(T) -> bool) -> bool

Typed version of Array::some.

Elements that can't be casted into T are effectively ignored (false is returned).

pub fn typed_iter<'a>(&'a self) -> impl Iterator<Item = T> + 'a

Typed version of Array::iter.

Elements that can't be casted into T are omitted.

#[must_use]pub fn typed_to_vec(&self) -> Vec<T>

Typed version of Array::to_vec.

Elements that can't be casted into T are omitted.

pub fn typed_keys(&self) -> impl Iterator<Item = u32>

Typed version of Array::keys.

pub fn typed_entries(&self) -> impl Iterator<Item = (u32, Result<T, JsValue>)>

Typed version of Array::entries.

Returns Err(JsValue)s for elements that could not be casted as T and elements that would have caused TypeErrors (i.e. non-objects). See js_sys::Iterator::next for more details about the latter case.

pub fn typed_values(&self) -> impl Iterator<Item = Result<T, JsValue>>

Typed version of Array::values.

Returns Err(JsValue)s for elements that could not be casted as T and elements that would have caused TypeErrors (i.e. non-objects). See js_sys::Iterator::next for more details about the latter case.

impl<T: JsCast> ReadOnlyArray<T>

#[must_use]pub fn mutable_array_copy(&self) -> Array

Produces a copy of the array that is mutable (i.e. an Array).

#[must_use]pub fn excluding_other_types(&self) -> Self

Produces a copy of the array that excludes any elements that can't be cast as T.

Indexing into this array (for valid indexes), for example, should always produce Some(_)s.