Struct xterm_js_sys::xterm::BufferCell

#[repr(transparent)]pub struct BufferCell {
    obj: JsValue,
}

Represents a single cell in the terminal’s buffer.

(This is a duck-typed interface).

Fields
obj: JsValue
Implementations

impl BufferCell

pub fn get_bg_color(&self) -> Color

Gets a cell’s background color number, this differs depending on what the color mode of the cell is:

Default: This should be 0, representing the default background color (CSI 49 m).

Palette: This is a number from 0 to 255 of ANSI colors (CSI 4(0-7) m, CSI 10(0-7) m, CSI 48 ; 5 ; 0-255 m).

RGB: A hex value representing a ‘true color’: 0xRRGGBB (CSI 4 8 ; 2 ; Pi ; Pr ; Pg ; Pb)

impl BufferCell

pub fn get_bg_color_mode(&self) -> u8

Gets the number representation of the background color mode, this can be used to perform quick comparisons of 2 cells to see if they’re the same. Use is_bg_rgb, is_bg_palette, and is_bg_default to check what color mode a cell is.

impl BufferCell

pub fn get_chars(&self) -> Str

The character(s) within the cell. Examples of what this can contain:

A normal width character

A wide character (eg. CJK)

An emoji

impl BufferCell

pub fn get_code(&self) -> u32

Gets the UTF32 codepoint of single characters, if content is a combined string it returns the codepoint of the last character in the string.

impl BufferCell

pub fn get_fg_color(&self) -> Color

Gets a cell’s foreground color number, this differs depending on what the color mode of the cell is:

Default: This should be 0, representing the default foreground color (CSI 39 m).

Palette: This is a number from 0 to 255 of ANSI colors (CSI 3(0-7) m, CSI 9(0-7) m, CSI 38 ; 5 ; 0-255 m).

RGB: A hex value representing a ‘true color’: 0xRRGGBB. (CSI 3 8 ; 2 ; Pi ; Pr ; Pg ; Pb)

impl BufferCell

pub fn get_fg_color_mode(&self) -> u8

Gets the number representation of the foreground color mode, this can be used to perform quick comparisons of 2 cells to see if they’re the same. Use is_fg_rgb, is_fg_palette, and is_fg_default to check what color mode a cell is.

impl BufferCell

pub fn get_width(&self) -> u8

The width of the character. Some examples:

1 for most cells.

2 for wide character like CJK glyphs.

0 for cells immediately following cells with a width of 2.

impl BufferCell

pub fn is_attribute_default(&self) -> bool

Whether the cell has the default attribute (no color or style).

impl BufferCell

pub fn is_bg_default(&self) -> bool

Whether the cell is using the default background color mode.

impl BufferCell

pub fn is_bg_palette(&self) -> bool

Whether the cell is using the palette background color mode.

impl BufferCell

pub fn is_bg_rgb(&self) -> bool

Whether the cell is using the RGB background color mode.

impl BufferCell

pub fn is_blink(&self) -> bool

Whether the cell has the inverse attribute (CSI 5 m).

impl BufferCell

pub fn is_bold(&self) -> bool

Whether the cell has the bold attribute (CSI 1 m).

impl BufferCell

pub fn is_dim(&self) -> bool

Whether the cell has the inverse attribute (CSI 2 m).

impl BufferCell

pub fn is_fg_default(&self) -> bool

Whether the cell is using the default foreground color mode.

impl BufferCell

pub fn is_fg_palette(&self) -> bool

Whether the cell is using the palette foreground color mode.

impl BufferCell

pub fn is_fg_rgb(&self) -> bool

Whether the cell is using the RGB foreground color mode.

impl BufferCell

pub fn is_inverse(&self) -> bool

Whether the cell has the inverse attribute (CSI 7 m).

impl BufferCell

pub fn is_invisible(&self) -> bool

Whether the cell has the inverse attribute (CSI 8 m).

impl BufferCell

pub fn is_italic(&self) -> bool

Whether the cell has the inverse attribute (CSI 3 m).

impl BufferCell

pub fn is_underline(&self) -> bool

Whether the cell has the underline attribute (CSI 4 m).