Struct xterm_js_sys::xterm::Terminal

#[repr(transparent)]pub struct Terminal {
    obj: Disposable,
}

The class that represents an xterm.js terminal.

Fields
obj: Disposable
Implementations

impl Terminal

pub fn new(options: Option<TerminalOptions>) -> Terminal

Creates a new Terminal object.

Takes options: an object containing a set of options.

impl Terminal

pub fn buffer(&self) -> BufferNamespace

[EXPERIMENTAL] The terminal’s current buffer, this might be either the normal buffer or the alt buffer depending on what’s running in the terminal.

impl Terminal

pub fn cols(&self) -> u16

The number of columns in the terminal’s viewport. Use TerminalOptions::cols to set this in the constructor and Terminal::resize for when the terminal exists.

impl Terminal

pub fn element(&self) -> Option<Element>

The element containing the terminal.

impl Terminal

pub fn markers(&self) -> ReadOnlyArray<Marker>

[EXPERIMENTAL] Get all markers registered against the buffer.

If the alt buffer is active this will always return [] (an empty array).

impl Terminal

pub fn on_binary(&self, listener: &Closure<dyn FnMut(Str)>) -> Disposable

Adds an event listener for when a binary event fires.

This is used to enable non UTF-8 conformant binary messages to be sent to the backend. Currently this is only used for a certain type of mouse reports that happen to be not UTF-8 compatible. The event value is a String, pass it to the underlying pty as binary data, e.g. pty.write(Buffer.from(data, 'binary')).

Returns a Disposable to stop listening.

See attach_binary_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_cursor_move(&self, listener: &Closure<dyn FnMut()>) -> Disposable

Adds an event listener for the cursor moves.

Returns a Disposable to stop listening.

See attach_cursor_move_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_data(&self, listener: &Closure<dyn FnMut(Str)>) -> Disposable

Adds an event listener for when a data event fires.

This happens, for example, when the user types or pastes into the terminal. The event value is whatever String results; in a typical setup, this should be passed on to the backing pty.

Returns a Disposable to stop listening.

See attach_data_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_key(&self, listener: &Closure<dyn FnMut(KeyEventData)>) -> Disposable

Adds an event listener for when a key is pressed.

The event value (KeyEventData) contains the string that will be sent in the data event as well as the DOM event that triggered it.

Returns a Disposable to stop listening.

See attach_key_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_line_feed(&self, listener: &Closure<dyn FnMut()>) -> Disposable

Adds an event listener for when a line feed is added.

Returns a Disposable to stop listening.

See attach_line_feed_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_render(
&self,
listener: &Closure<dyn FnMut(RenderEventData)>
) -> Disposable

Adds an event listener for when rows are rendered.

The event value (RenderEventData) contains the start row and end row of the rendered area (ranges from 0 to [Terminal::rows] - 1).

Returns a Disposable to stop listening.

See attach_render_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_resize(
&self,
listener: &Closure<dyn FnMut(ResizeEventData)>
) -> Disposable

Adds an event listener for when the terminal is resized.

The event value (ResizeEventData) contains the new size.

Returns a Disposable to stop listening.

See attach_resize_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_scroll(&self, listener: &Closure<dyn FnMut(u32)>) -> Disposable

Adds an event listener for when a event listener for when a scroll occurs.

The event value is the new position of the viewport.

Returns a Disposable to stop listening.

See attach_scroll_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_selection_change(&self, listener: &Closure<dyn FnMut()>) -> Disposable

Adds an event listener for when a selection change occurs.

Returns a Disposable to stop listening.

See attach_selection_change_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn on_title_change(&self, listener: &Closure<dyn FnMut(Str)>) -> Disposable

Adds an event listener for when an OSC 0 or OSC 2 title change occurs.

The event value is the new title.

Returns a Disposable to stop listening.

See attach_title_change_event_listener (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn parser(&self) -> Parser

[EXPERIMENTAL] Get the parser interface to register custom escape sequence handlers.

impl Terminal

pub fn rows(&self) -> u16

The number of rows in the terminal’s viewport. Use TerminalOptions.rows to set this in the constructor and Terminal::resize for when the terminal exists.

impl Terminal

pub fn textarea(&self) -> Option<HtmlTextAreaElement>

The textarea that accepts input for the terminal.

impl Terminal

pub fn unicode(&self) -> UnicodeHandling

[EXPERIMENTAL] Get the Unicode handling interface.

This can be used to register Unicode versions and switch the active Unicode version.

impl Terminal

pub fn strings(&self) -> LocalizableStrings

Natural language strings that can be localized.

impl Terminal

pub fn blur(&self)

Unfocus the terminal.

impl Terminal

pub fn clear(&self)

Clear the entire buffer, making the prompt line the new first line.

impl Terminal

pub fn clear_selection(&self)

Clears the current terminal selection.

impl Terminal

pub fn focus(&self)

Focus the terminal.

impl Terminal

pub fn get_selection(&self) -> Str

Gets the terminal’s current selection; this is useful for implementing copy behavior outside of xterm.js.

impl Terminal

pub fn get_selection_position(&self) -> Option<SelectionPosition>

Gets the selection position or None if there is no selection.

impl Terminal

pub fn has_selection(&self) -> bool

Gets whether the terminal has an active selection.

impl Terminal

pub fn load_addon(&self, addon: TerminalAddon)

Loads an addon into this instance of the xterm.js Terminal.

Takes:

addon: The addon to load.

See load_xterm_addon (if the ext feature is enabled) for a friendlier version of this function.

impl Terminal

pub fn open(&self, parent: Element)

Opens the terminal within an element.

Takes:

parent: The element to create the terminal within. This element must be visible (have dimensions) when open is called as several DOM-based measurements need to be performed when this function is called.

impl Terminal

pub fn paste(&self, data: Str)

Writes text to the terminal, performing the necessary transformations for pasted text.

Takes:

data: The text to write to the terminal.

impl Terminal

pub fn refresh(&self, start: u16, end: u16)

Tells the renderer to refresh terminal content between two rows (inclusive) at the next opportunity.

Takes:

start: The row to start from (between 0 and [Terminal::rows] - 1).

end: The row to end at (between start and [Terminal::rows] - 1).

impl Terminal

pub fn reset(&self)

Perform a full reset (RIS, aka ‘\x1bc’).

impl Terminal

pub fn resize(&self, columns: u16, rows: u16)

Resizes the terminal.

It’s best practice to debounce calls to resize, this will help ensure that the pty can respond to the resize event before another one occurs.

impl Terminal

pub fn scroll_lines(&self, amount: i32)

Scroll the display of the terminal.

Takes:

amount: The number of lines to scroll down (negative scrolls up).

impl Terminal

pub fn scroll_pages(&self, page_count: i32)

Scroll the display of the terminal by a number of pages.

Takes:

page_count: The number of pages to scroll (negative scrolls up).

impl Terminal

pub fn scroll_to_bottom(&self)

Scrolls the display of the terminal to the bottom.

impl Terminal

pub fn scroll_to_line(&self, line: u32)

Scrolls to a line within the buffer.

Takes:

line: The 0-based line index to scroll to.

impl Terminal

pub fn scroll_to_top(&self)

Scrolls the display of the terminal to the top.

impl Terminal

pub fn select(&self, column: u16, row: u32, length: u32)

Selects text within the terminal.

Takes:

column: The column the selection starts at.

row: The row the selection starts at.

length: The length of the selection.

impl Terminal

pub fn select_all(&self)

Selects all text within the terminal.

impl Terminal

pub fn select_lines(&self, start: u32, end: u32)

Selects text in the buffer between 2 lines.

Takes:

start: The 0-based line index to select from (inclusive).

end: The 0-based line index to select to (inclusive).

impl Terminal

pub fn write(&self, data: Str)

Writes data to the terminal.

Takes:

data: The data to write to the terminal. The actual API allows for this to be either raw bytes given as Uint8Array from the pty or a string (raw bytes will always be treated as UTF-8 encoded, string data as UTF-16). For simplicity, we just take a String; this shouldn't cause problems (going from UTF-8 encoded Rust Strings to UTF-16 JS strings) and just makes things simpler.

impl Terminal

pub fn write_with_callback(&self, data: Str, callback: &Closure<dyn FnMut()>)

Writes data to the terminal and takes a callback.

This identical to write except it also takes a callback.

Takes:

data: The data to write to the terminal. The actual API allows for this to be either raw bytes given as Uint8Array from the pty or a string (raw bytes will always be treated as UTF-8 encoded, string data as UTF-16). For simplicity, we just take a String; this shouldn't cause problems (going from UTF-8 encoded Rust Strings to UTF-16 JS strings) and just makes things simpler.

callback: Callback that fires when the data was processed by the parser.

impl Terminal

pub fn load_xterm_addon<A: IntoJsInterface<TerminalAddon>>(&self, addon: &A)

This is supported on feature="ext" only.

Loads an addon into this instance of the xterm.js Terminal.

This supports Rust defined addons as well as JS defined addons and is otherwise identical to Terminal::load_addon.

impl Terminal

#[must_use]pub fn new_with_wrapper(
options: Option<TerminalOptions>
) -> DisposableWrapper<Terminal>

This is supported on feature="ext" only.

Terminal constructor that encloses the resulting Terminal in a DisposableWrapper.

This is otherwise identical to Terminal::new.

impl Terminal

pub fn attach_binary_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(Str),
F: 'static,

This is supported on feature="ext" only.

Attaches a binary event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_binary.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_cursor_move_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(),
F: 'static,

This is supported on feature="ext" only.

Attaches a cursor move event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_cursor_move.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_data_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(Str),
F: 'static,

This is supported on feature="ext" only.

Attaches a data event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_data.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_key_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(KeyEventData),
F: 'static,

This is supported on feature="ext" only.

Attaches a key event event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_key.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_line_feed_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(),
F: 'static,

This is supported on feature="ext" only.

Attaches a line feed event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_line_feed.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_render_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(RenderEventData),
F: 'static,

This is supported on feature="ext" only.

Attaches a render event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_render.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_resize_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(ResizeEventData),
F: 'static,

This is supported on feature="ext" only.

Attaches a resize event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_resize.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_scroll_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(u32),
F: 'static,

This is supported on feature="ext" only.

Attaches a scroll event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_scroll.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_selection_change_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(),
F: 'static,

This is supported on feature="ext" only.

Attaches a selection change event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_selection_change.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!

pub fn attach_title_change_event_listener<F>(
&self,
listener: F
) -> DisposableWrapper<Disposable> where
F: FnMut(Str),
F: 'static,

This is supported on feature="ext" only.

Attaches a title change event listener and returns a DisposableWrapper that can be dropped to make xterm.js stop sending the event listener events. This is sugar for Terminal::on_title_change.

We assume event listener closures are going to be long-lived, so we leak the closure that is produced here!